Create the new project.
Refinery is a replacement for xls2yml, and will take in CSVs, create a
Turbine graph, and output YAML for storage in ETsource.

Prior to outputting the YAML, the Turbine graph will be modified by
"catalysts" -- this may be removing slot conversions for loss output
slots, performing demand calculations, or any number of other features
we extract from InputExcel.

Refinery is part of the InputExcel replacement project.