Add demand / share calculators.
The previous approach of doing calculation inside the catalysts wasn't
very sustainable long-term; as more complicated arrangements of nodes
and edges were added, it became hard to add additional ways for
calculating shares and demands.

The calculators, one of which is assigned to every node and edge, will
be looped through until the whole graph has been calculated
(ETengine-style).