Add Demand::FillRemaining strategy.
Fixes test failures introduced in a recent commit. This attempts to
assign demand to a node by looking at unfulfilled demand on it's
children, and assigning that to the node. This only works if the
children have no other parents, or if we already know the expected
demand of these parents.