Add FromDemand edge share calculation strategy.
A strategy for calculating the share of an edge by looking at the
demand of the parent and child nodes. It correctly accounts for
demand which may be provided to the child nodes by *other*, related
parents.

In the example graph below, FromDemand is capable of calculating the
shares of the [M]->[S] and [M]->[C] edges.

               (100)
    (5) [R]     [M]     [F] (100)
          \     / \     /
     (1.0) \   /   \   / (1.0)
            \ /     \ /
       (80) [S]     [C] (125)