Nothing to see here.
Okay, okay: "Edge shares should be a share of the input".

I had misinterpreted the original requirements -- and, for that matter,
what link shares are -- and believed that the share was the proportion
of energy from an output. For example, given the following graph:

  (5) [A] [B] (5)
        \ /
        [X] (10)

... the edge shares were each calculated to be 1.0; each node [A] and
[B] output all their energy to the child [X].

NOPE!

In fact, edge shares are the proportion of the *input*; in the above
example each edge should have a share of 0.5, since they supply half of
[X]'s energy. In essence, Refinery had been calculating the "technology
share" of each edge.

Edge share calculators have been turned inside out (in the process,
losing my keys and part of my sanity) to correctly calculate the share
of each edge.

Following this, I will seek to replace edge share calculations with an
edge *demand* calculation (from which we can easily determine share
later) since this would be the same from the perspective of the parent
and child nodes.

Yikes.