Allow slot demands to vary by 1e-20 from the node.
Very slightly relaxes validation of slots so that their demands may vary
by a *tiny* amount from the demand of the node. In ETSource, numbers are
often expressed in petajoules, so this variation is equivalent to
0.00001 joules.

Or to put it another way, we allow slots to be inaccurate by the same
amount as a human body uses in 0.0001 seconds.