Try to calculate slot shares automatically.
Whenever the slot is the only slot on the node, or we have other
information available (the slot demand, and the demand of the node or
sibling slots) we can calculate the share automatically.

This allows Refinery to fill in the slot share without it having to be
specified by a user.