Add support for ETEngine-style flexible edges
Flexible edges supply energy to a node to ensure that it fulfils demand.
For example, given a child and two parents, if the child demand 100, and
one parent supplies 50, then the other parent must also supply 50.

This would ordinarily be handled by the FillRemaining strategy, except
when the node also has an uncalculated "overflow" edge, in which case an
incoming flexible edge will often be required.