Add a "deeper" node demand calculation
This alteration allows us to calculate a node demand when we know only a
single edge demand, but can augment that with an appropriate edge share
and slot share.

This adds to the existing strategy which did not require an edge share,
but requires all the slot's edges to have a demand.

Ref #31, quintel/rdr#64.