Make FromDemand work across carriers; add FromChildDemand
Fixes assignment of remaining demand / supply to or from a node even
when the edges are in different carriers.

Added FromChildDemand which works in the opposite direction. I consider
this to be a temporary addition until I come up with something more
flexible which allows strategies to be written in a "generic" way, so
that Refinery can automatically run them in both directions.