Add support for reversible strategies
A reversible strategy is written using helper methods (in_edges,
parent_share), etc, and is then able to be executed in the opposite
direction.

This eliminates the need to effectively implement (almost) every
strategy twice, once for each direction, in favour of having Refinery
do it automatically.

A reversible strategy mixes in the "Reversible" module and is then
"compile"d to the chosen direction.

    class SomeStrategy
      include Reversible
      # ...
    end

    SomeStrategy.compile(:forwards).new