Recursive max demand should not stop at the first traversal
Ensures that calculating max demand recursively will actually recurse
until it reaches a node with a max_demand, instead of stopping as soon
as it peforms a single traversal.

Each edge through which it should traverse MUST have a priority to make
it behave as a flex-max edge, instead of a normal flexible edge.