Improve validation error messages
Mismatches of demand now show the unit, and will adjust the number so
that a worthwhile value can be shown. For example, when an expected and
actual demand are extremely similiar, the difference between them will
be shown, converted to a smaller unit size in order to now show a
difference of zero.

e.g.

    "difference of 0.00000 PJ"
    "difference of 0.03 MJ"