Allow small deviations in max_demand validation
As max_demand is not a precise property (it cannot be reliably cast to a
Rational as the value may be Float::INFINITY or "recursive") floating
point errors may creep in while calculating max demand.

This can cause erroneous failures when validating that node demand does
not exceed max demand.

Ref quintel/etsource#1809