Fix flexible edges ignoring sibling share
Fix some cases where the calculation of a forward reversed edge would
ignore the share of siblings. This was due to incorrect calculation of
parent demand which would look for an edge which was the only input to
the "to" node, but then wrongly took demand from the "from" node.

Closes #60